Add prototype of React Component displaying configurations source types
ADR-0011 introduces the API for managing global configurations. Thus, the
Trace Extension needs an interface to expose this API to users.

The purpose of this commit is to start creating this UI for the trace
extension as a React Component. It provides a skeleton for the component
so that it can be further extended in future commits:

[1] Added TraceConfigurationManager to handle calls to the TSPClient that
relates to trace configuration.

[2] Introduce a prototype of the UI as the
TraceConfigurationsDialogComponent. This is the main component of the UI
that handles the display logic for other sub-components, such as [3].

[3] Added TraceConfigurationListComponent to list the available
configuration source types and their instances from the server.

To test, open a trace and hover the mouse over the toolbar of the
Available Views widget. There should be an icon that appears. Click on the
icon to open the trace configuration dialog. In this prototype, the dialog
should only display the configuration source types.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>