[ci] Use GitHub token to avoid ripgrep download issue
In a recent CI run, I noticed this error:
"Downloading ripgrep failed: Error: Request failed: 403"

This happens because of GitHub has limits when anonymously accessing
its API, which we do when downloading ripgrep. There's an easy fix:
use an auto-generated GitHub token when building the Theia app.

Also tidied-up the workflow file a little, to make it easier to read.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>