update following upstream Theia changes
There were some recent Theia changes[1] that required to change the way our
Theia extension starts. It used to be that our TraceServerUrlProviderImpl
backend contribution could wait in the "initialize" Theia app life-cycle
phase, until the application was started and it received the trace server
port from the front-end, from our PreferencesFrontendContribution.

But the above no longer works. Now backend contributions are started in
parallel, and all are expected to finish their initialize phase, before the
"start" phase is initiated.

In practice, after upgrading to Theia v1.45.1, the Theia application, that
includes the trace viewer extension,  would be killed during startup, with
no clear indication why.

To prevent this from happening, the need to receive information from the
front-end, during initialization, has been removed: the backend contribution
now initializes with the assumption that the default trace server port is used.
If a different port is configured in the preferences, the information will be
synched after the app has started.

[1]: https://github.com/eclipse-theia/theia/pull/12818

specifically:

"The BackendApplicatinContribution init and configure methods are now called
in parallel for all contributions instead of sequentially."

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>