Update release instructions
During the recent first automated release, we noticed that there's an
extra step to be done in this repository. It's because the version of
"theia-traceviewer" needs to be updated in the example applications.

I think this does not happen automatically, even though the example
applications are part of the lerna workspace, because of our current
.gitignore file, which ignores these folders.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>