[examples] update .gitignore strategy: blanket ignore -> granular ignore
ATM the .gitignore is configured to blanket ignore the examples, with a few
punctual exceptions. This has the non-obvious side-affect with "lerna version",
which then ignores these folders even if they are part of the lerna workspace.
This results in our root package.json "version:<*>" entries not updating the
version of the "theia-traceviewer" dependency in the examples' package.json,
requiring this be done manually.

The commit implements the opposite strategy for the examples: only ignore the
files/ folders that should not be checked-in.

See below for a test run of "yarn version:major" that shows that, with this
change,  the examples applications are now handled automatically:

theia-trace-extension$ yarn version:major
yarn run v1.22.21
$ lerna version major --exact --no-push --git-tag-command /usr/bin/true --yes -m "Release %s (Major)"
lerna notice cli v7.4.2
lerna info current version 0.2.0
lerna info Looking for changed packages since v0.1.0

Changes:
 - browser-theia-trace-example-app: 0.1.0 => 1.0.0 (private)
 - electron-theia-trace-example-app: 0.1.0 => 1.0.0 (private)
 - traceviewer-base: 0.2.0 => 1.0.0
 - traceviewer-react-components: 0.2.0 => 1.0.0
 - theia-traceviewer: 0.2.0 => 1.0.0

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>