Streamline Theia example applications, making them RCP-like
Now that there exists is a full-fledge Theia IDE application [1]
that incorporates the theia-trace-viewer, the example applications
(browser, electron) in this repo, need not be more than "trace viewer
appliances", that are used for testing purpose.

This commit removes most Theia framework extensions for the two
example applications, transforming them into something relatively
close to appliances.

There remains a few remnants like the outline and problems views, that
are still pulled by essential Theia framework extension that we need.

[1] cdt-cloud blueprint :
https://github.com/eclipse-cdt-cloud/cdt-cloud-blueprint

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>