Update view range during trace indexing
While the trace context component updateTrace() loop is iterating,
update the unit controller's view range if the current view range is the
full range and the full range has changed.

If the user zooms the view range, it will stop updating to prevent
snapping out of the user's view range. If the users zooms back out to
the full range it will restart updating again.

After updateTrace() is complete, update the view range only if the
current view range is the full range.

In TimegraphOutputComponent, use the analysis-running-overflow overlay
while the analysis is running.

Let the analysis-running-overflow pass through pointer events to the
underlying layer.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>