Adapt to later ag-grid version
Dependabot opened a PR to suggest we update our "ag-grid" dependencies.
However, the PR did not build, because ir proposes we jump forward 3
major releases forward, and there were API changes between that version
and what we use.

In consequence, it's necessary to make changes in how we import and use this
dependency. Here are a few highlights:

- Switch to use the "@ag-grid-community/*" modules
- Theme ag-theme-balham.css: both light and dark themes are now bundled in
  the same file
- ColumnApi has been deprecated, then removed. its features have been moved
  to the GridApi, which we now need to use
- need to explicitly import and register infinte-row-model module, before we can
  use it to scroll in our tables: @ag-grid-community/infinite-row-model
- no longer possible to set data source using "api.setDatasource([...])". now
  using "api.setGridOption('datasource', [...])" instead

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>