Re-enable all React table-renderer-components tests
Fixes #1100

Had to mock some ag-grid objects that were used in the newly re-enabled tests.

It turns-out that many of these objects were not really used during our tests
(e.g. api: GridApi, column: Column). So for these, it was enough to mock
them using an empty JS object and use keyword "as" to present them as the real
thing. e.g.:

<SearchFilterRenderer [...]
    api={{} as GridApi}
    column={{} as Column<any>}
[...]>

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>