Create a new 'Open in Trace Viewer' command, add it to navigator context menu
Starting around Theia v1.50.0, the "open-with" service no longer permitted to open
folders from the navigator context menu (only worked for files). It was possible to
make it work again, but there were undesirable side-effects. So instead, we decided
to register our own command, "Open in Trace Viewer", and add it to the navigator's
context menu at the first level. This is very similar to what we have on the `Trace
Viewer for VSCode` extension.

Also updated the README to reflect how a trace can now be be opened from the File
Explorer, and recorded a new .gif showing it.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>