update Electron dependencies, pin nan@2.18.0
Also ran yarn upgrade, to also update the other dependencies in the repo that
might benefit, following the update of the Electron-related dependencies.

With recent Theia, the rebuilding of the native dependencies for Electron
does not work unless dependency "nan" is pinned to v2.18.0. For that reason,
that dependency is added to this repo's root package.json's resolutions block.

This issue highlighted that we do not perform a full build of the Electron app
by default, leaving the rebuilding of the native dependencies (theia rebuild ...)
to be done before starting the Electron app or packaging it. In consequence,
it took a while to notice there was a problem with Electron once we upgraded to
Theia v1.52.0.

Also, update Jest test snapshots - there is a slight difference in the generated
HTML code for one testcase, following the minor ag-grid update in this commit.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>