Fix time graph filtering and collapsing
Update the filtering of empty nodes to include visible nodes even if any
of their parent is filtered-out. Recursively replace filtered-out nodes
by their visible children, if any, but only if the filtered-out parent
node is not collapsed.

In TimeGraphOutputComponent update the arrows layer after empty nodes
have been determined, and filter out empty nodes from the rowIds passed
to the arrows layer.

Update getIndexOfNode() in utils to consider the list of empty nodes.

Fix names of selectedRow and multiSelectedRows in TableBodyProps.

Fixes #1132

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>