Support of directory traces other than CTF
The trace detection algorithm checks recursively for CTF traces. If no
CTF trace are found under the root directory, then use the root
directory as trace directory. The trace server back-end will validate
if this path points to a valid trace or not.

With this it's now possible to open traces that are directories other
than CTF.

Note that only a single trace can be opened in the experiment and not
set of traces, which is currently only supported with CTF traces.

Examples trace type: uftrace

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>