Enable reset of entry order in entry trees
Add onSortReset callback to table, and add onOrderReset callbacks to
tree and entry tree.

Trigger the onSortReset callback from table when the first column is
cycled back to the default sort order.

Store the default ordered ids when fetching the tree in timgraph output
component, xy output component and datatree output component. When the
onOrderReset callback is handled, trigger the ordering of the tree using
the stored default order.

Remove the unused orderedNodes from the state of xy output component and
datatree output component.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>