Update for Theia community release v1.58.4
Issue encountered:
https://github.com/eclipse-theia/theia/issues/14804

A PR updated "fix-path" and caused the issue:
https://github.com/eclipse-theia/theia/pull/14781

A fix was done:
https://github.com/eclipse-theia/theia/pull/14812

Side-effect of the fix:
https://github.com/eclipse-theia/theia/issues/14804#issuecomment-2630660868

Fix to address the side-effect:
https://github.com/eclipse-theia/theia/pull/14819

Note: the fix for the side-effect was merged on Theia master,
and does not help momentariry on the community release 1.58.x,
on which it's not currently planned to back-port this fix:
https://github.com/eclipse-theia/theia/pull/14819#issuecomment-2786415441

We could contribute a PR to back-port the fix, but I do not
think this is super important for us - the Theia Electron app
is only an example application and permanently switching to
a bundled Electron backend does not seem to have any downside.

In consequence, this PR includes using the bundled Electron
backend, which is a simple webpack config update. If we later
notice issues with this, we can roll-back to use the non-bundled
backend in the next community release, which will have the fix
above.

Note about the webpack config update: as suggested, I have
deleted the webpack configs and run "theia build" to have
them re-created. Then I re-added the small change we need
for the time range context menu. The result is a webpack
config that's more aligned with the one Theia generates,
for both our example Theia apps.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>