update to theia community release 1.61.0
As part of the update, add the context to the object passed to
contextMenuRenderer.render(). This fixes the new compile errors,
in 3 places.

This change is needed to adapt to the following breaking Theia
upstream change:
https://github.com/eclipse-theia/theia/pull/14982

Also, had to update @theia/playwirght and related dependencies for
the playwright test suite, used in CI. All tests are now passing
again.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>