[customized views] remove UIs related to managing them
In the Theia trace extension, the functionality to create custom
views has not been implemented. However the UI element to create
them is defined in the react components, used here too.

This commit disables the UI (button) to add custom views, displayed
on view entries that support it (in-and-out currently). For consistency,
the UI to remove custom views is also removed here, though it works
without additional code.

The above may seem strange, but it's possible, when sharing a trace
server, to create custom views using the vscode extension, and then
see them using the Theia extension.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>