[license check] update exclusions
Dependency "npm/npmjs/-/allure-commandline/2.34.0" was added
to the ignore list. It's used as a dev dependency and
believed to be license-compatible with the project, but it's
still being investigated by the IP team (see ticket URL in
exclusion file).

Also, took the opportunity to remove old exclusions that,
as per the output of the lwciense check, had no effect (they
were all approved since being ignored).

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>