doc: Archive graphql-server-prototype to remove it from project's SBOM
The SBOM of GitHub Insight will include the dependencies of the
graphql-server-prototype as dependency of the theia-trace-extension.
However, those should not be listed because they are not deployed and
used in production.

This PR archives the protoype code in a tar.gz file and updates the
instructions in the corresponding ADR to use it.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>